% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_api_key.R
\name{ws_construct_api_key}
\alias{ws_construct_api_key}
\alias{ws_api_key}
\alias{ws_api_key_header}
\alias{ws_list_apis}
\alias{ws_auth_header}
\title{Check API Keys for Various Web of Science APIs}
\usage{
ws_construct_api_key(api_key = NULL, error = TRUE, key_option, sys_env,
  key_filename_option, api_name = "")

ws_api_key(api_key = NULL, error = TRUE, api = names(ws_list_apis()))

ws_api_key_header(api_key = NULL)

ws_list_apis()

ws_auth_header(api_key = NULL)
}
\arguments{
\item{api_key}{Web of Science API key}

\item{error}{Should the function error if \code{api_key = NULL} after
looking in various places?  (See notes)}

\item{key_option}{option to grab API key using \code{\link{getOption}}}

\item{sys_env}{System environment variable to grab API key from
\code{\link{Sys.getenv}}}

\item{key_filename_option}{Option retrieved from \code{\link{getOption}},
which is a location of a file with the API key.}

\item{api_name}{Name of API key}

\item{api}{an endpoint for the API, see
\code{\link{ws_list_apis}}}
}
\value{
API key character string
}
\description{
Check API Keys for Various Web of Science APIs
}
\note{
You can either set the API key using
\code{option(ws_API_key)} or have it accessible by
\code{api_key = Sys.getenv('API_KEY')}, where \code{API}
is one of the Web of Science APIs, such as \code{INCITES}
}
\examples{
api_key = NULL
error = FALSE
  api_key = ws_construct_api_key(
  api_key = api_key,
  error = error,
  key_option = "ws_incites_key",
  sys_env = "INCITES_KEY",
  key_filename_option = "ws_incites_key_filename",
  api_name = "Incites")
ws_list_apis()
ws_auth_header(api_key = "my_api_key")
}
