% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_build_request.R
\name{ws_build_request}
\alias{ws_build_request}
\title{Build Request for Web of Science}
\usage{
ws_build_request(api_key = NULL, api, endpoint, ...)
}
\arguments{
\item{api_key}{API key for the}

\item{api}{API to be used}

\item{endpoint}{endpoint for that specific API}

\item{...}{additional arguments to send to \code{\link{VERB}}}
}
\value{
A request output
}
\description{
Build Request for Web of Science
}
\examples{
ws_build_request(api_key = ws_incites_key(error = FALSE),
api = "incites", endpoint = "InCitesLastUpdated")
}
