% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_databases.R
\name{ws_databases}
\alias{ws_databases}
\alias{ws_db}
\title{Web of Science Databases}
\usage{
ws_databases()

ws_db()
}
\value{
character vector
}
\description{
Web of Science Databases
}
\examples{
ws_databases()
ws_db()
}
