% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endnote.R, R/incites.R, R/ws_endpoints.R
\name{ws_endnote_endpoints}
\alias{ws_endnote_endpoints}
\alias{ws_incites_endpoints}
\alias{ws_endpoints}
\alias{ws_woslite_endpoints}
\alias{ws_wos_endpoints}
\alias{ws_tipms_endpoints}
\title{Web of Science Endpoints}
\usage{
ws_endnote_endpoints()

ws_incites_endpoints()

ws_endpoints(out_type = c("json", "xml"), api = NULL)

ws_woslite_endpoints()

ws_wos_endpoints()

ws_tipms_endpoints()
}
\arguments{
\item{out_type}{output type, either \code{json} or \code{json}}

\item{api}{The API being used}
}
\value{
A list of endpoints for each API
}
\description{
Web of Science Endpoints
}
\examples{
ws_incites_endpoints()
ws_endpoints(api = "incites")
}
