% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_incites_documents_by_id.R
\name{ws_incites_by_inst_id}
\alias{ws_incites_by_inst_id}
\alias{ws_incites_by_record_count}
\alias{ws_incites_by_ut}
\alias{ws_incites_invalid_ut}
\alias{ws_incites_last_updated}
\title{Get Incites Documents by Institution ID}
\usage{
ws_incites_by_inst_id(year = 2018, record_count = 100,
  starting_record = 1, esci = c("n", "y"), api_key = NULL, ...)

ws_incites_by_record_count(year = 2018, api_key = NULL, ...)

ws_incites_by_ut(ut, esci = c("n", "y"), api_key = NULL, ...)

ws_incites_invalid_ut(ut, api_key = NULL, ...)

ws_incites_last_updated(api_key = NULL, ...)
}
\arguments{
\item{year}{Your response will contain records published
starting in this year. Must be at least 1980}

\item{record_count}{This is the maximum number of records that
will be returned per request. Must be less than the
limit placed on the developer key (default maximum is 100)}

\item{starting_record}{Starting index of Records returned.
Used when number of Records required exceeds limit
on single request. Starts at Index 1}

\item{esci}{Flag to return InCites baseline metrics with
Emerging Sources Citation Index (ESCI) content.
Accepts y/n case insensitive values only.
y - returns InCites metrics with ESCI index,
n - returns InCites metrics without ESCI index}

\item{api_key}{InCites Web of Science API key}

\item{...}{additional arguments to send to \code{\link{fromJSON}}}

\item{ut}{a vector of UT values and get InCites metrics in
the response. A valid UT is 15 alphanumeric digits.
All trailing 0’s must be left in the UT, and there
should be no spaces between UTs. If there is no data
for a UT or an invalid UT is entered, that UT will be
ignored in the response. A limit of 100 UTs may be submitted
per request. To see sample API results, use one or more
 of the following UTs}
}
\value{
A list of the request and the content
}
\description{
Get Incites Documents by Institution ID
}
\examples{
if (ws_have_incites_key()) {
result = ws_incites_by_inst_id(year = current_year(),
flatten = TRUE)
docs = result$content$api$rval[[1]]
result_with_esci = ws_incites_by_inst_id(year = current_year(),
flatten = TRUE, esci = "y")
esci_docs = result_with_esci$content$api$rval[[1]]
}
if (ws_have_incites_key()) {
rec = ws_incites_by_record_count(year = current_year(),
flatten = TRUE)
rec$content$api$rval
}
ut = c("000352040700014","000353267900023",
 "000346982900018","000346342200004")
if (ws_have_incites_key()) {
result = ws_incites_by_ut(ut = ut,
flatten = TRUE)
docs = result$content$api$rval[[1]]
}
ut = c("000352040700014","000353267900023",
 "000346982900018","000346342200004")
if (ws_have_incites_key()) {
result = ws_incites_invalid_ut(ut = ut,
flatten = TRUE)
docs = result$content$api$rval[[1]]
}
if (ws_have_incites_key()) {
result = ws_incites_last_updated(flatten = TRUE)
result$content$api$rval[[1]]
}
}
