% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incites.R
\name{ws_incites_key}
\alias{ws_incites_key}
\alias{ws_have_incites_key}
\title{Find API Key for Incites}
\usage{
ws_incites_key(api_key = NULL, error = TRUE)

ws_have_incites_key(api_key = NULL)
}
\arguments{
\item{api_key}{Web of Science API key}

\item{error}{Should the function error if \code{api_key = NULL} after
looking in various places?  (See notes)}
}
\description{
Determines if \code{option(ws_incites_key)} or
\code{option(ws_incites_key_filename)} is set.
If not, it stops and returns an error.  If so, returns the value.
}
\examples{
res = ws_incites_key(error = FALSE)
tfile = tempfile()
writeLines("my_api_key", con = tfile)
options(ws_incites_key_filename = tfile)
ws_incites_key()
file.remove(tfile)
options(ws_incites_key_filename = NULL)

ws_filename = getOption("ws_incites_key_filename")
INCITES_KEY = Sys.getenv("INCITES_KEY")
Sys.setenv(INCITES_KEY = "")
writeLines(c("my_api_key1", "api_key2"), con = tfile)
options(ws_incites_key_filename = tfile)
testthat::expect_warning(ws_incites_key())
file.remove(tfile)
Sys.setenv(INCITES_KEY = INCITES_KEY)
options(ws_incites_key_filename = ws_filename)
ws_have_incites_key()
}
