% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_request_content.R
\name{ws_request_content}
\alias{ws_request_content}
\title{Get Content from Request}
\usage{
ws_request_content(request, error = FALSE, ...)
}
\arguments{
\item{request}{Request from \code{\link{VERB}}}

\item{error}{Should errors (compared to warnings)
be issued}

\item{...}{additional arguments to send to \code{\link{fromJSON}}}
}
\value{
An list from \code{\link{jsonlite}}
}
\description{
Get Content from Request
}
