% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endnote.R, R/incites.R, R/tipms.R, R/wos.R,
%   R/woslite.R, R/ws_api_key.R
\name{ws_set_endnote_key}
\alias{ws_set_endnote_key}
\alias{ws_set_incites_key}
\alias{ws_set_tipms_key}
\alias{ws_set_wos_key}
\alias{ws_set_woslite_key}
\alias{ws_set_key_option}
\title{Set API Key for API endpoint}
\usage{
ws_set_endnote_key(api_key)

ws_set_incites_key(api_key)

ws_set_tipms_key(api_key)

ws_set_wos_key(api_key)

ws_set_woslite_key(api_key)

ws_set_key_option(api_key, key_option)
}
\arguments{
\item{api_key}{Web of Science API key for specific endpoint}

\item{key_option}{option to set API key using \code{\link{options}}}
}
\description{
Sets API key using
if \code{options(key_option)}
}
\examples{
ws_endnote_key = getOption("ws_endnote_key")
ws_set_endnote_key(ws_endnote_key)
ws_set_incites_key("my_incites_api_key")
ws_tipms_key = getOption("ws_tipms_key")
ws_set_tipms_key(ws_tipms_key)
ws_wos_key = getOption("ws_wos_key")
ws_set_wos_key(ws_wos_key)
ws_woslite_key = getOption("ws_woslite_key")
ws_set_woslite_key(ws_woslite_key)
ws_set_key_option(api_key = "my_api_key", key_option = "ws_incites_key")
}
