% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woslite.R
\name{ws_woslite_key}
\alias{ws_woslite_key}
\alias{ws_have_woslite_key}
\title{Find API Key for WOS Lite}
\usage{
ws_woslite_key(api_key = NULL, error = TRUE)

ws_have_woslite_key(api_key = NULL)
}
\arguments{
\item{api_key}{Web of Science API key}

\item{error}{Should the function error if \code{api_key = NULL} after
looking in various places?  (See notes)}
}
\description{
Determines if \code{option(ws_woslite_key)} or
\code{option(ws_woslite_key_filename)} is set.
If not, it stops and returns an error.  If so, returns the value.
}
\examples{
res = ws_woslite_key(error = FALSE)
ws_have_woslite_key()
}
